# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "MetaShare Api",
    "description": "MetaShare 相關 Api"
  },
  "host": "127.0.0.1:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions":{
    "Bearer":{
      "type": "apiKey",
      "in": "headers",
      "name": "Authorization",
      "description": "JWT Token"
    }
  },
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/api/post/{id}/likes": {
      "get": {
        "tags": ["Post"],
        "description": "取得單則貼文喜歡的人",
        "security":[{
          "Bearer":[]
        }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "成功回應",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "array",
                  "example": [ {
                    "_id": "627100547ef69b72689e67f6",
                    "name": "Chia Yu",
                    "avator": "https://imgur.com/LwPqBcI.png"
                }]
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400 (ID 格式錯誤)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/idFormatErrorResponse"
            }
          },
          "400 (查無此 ID)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/notFoundIdErrorResponse"
            }
          },
          "400": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/errorResponse"
            }
          }
        }
      },
      "patch": {
        "tags": ["Post"],
        "description": "喜歡/收回喜歡",
        "security":[{
          "Bearer":[]
        }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "成功回應",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "array",
                  "example": [{
                    "_id": "62763cb70b6e444e6fd4fbbb",
                    "author": {
                      "_id": "627100547ef69b72689e67f6",
                      "name": "Chia Yu",
                      "avator": "https://imgur.com/LwPqBcI.png"
                    },
                    "content": "更新內容",
                    "imageUrls": [],
                    "likes": [
                      "627100547ef69b72689e67f6"
                    ],
                    "createdAt": "2022-05-07T09:32:39.272Z",
                    "comments": []
                  }]
                },
                "message": {
                  "type": "string",
                  "example": "更新成功"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400 (ID 格式錯誤)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/idFormatErrorResponse"
            }
          },
          "400 (查無此 ID)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/notFoundIdErrorResponse"
            }
          },
          "400 (尚未登入)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/userIdErrorResponse"
            }
          },
          "400": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/errorResponse"
            }
          }
        }
      }
    },
    "/api/post/{id}/comments": {
      "get": {
        "tags": ["Post"],
        "description": "取得單則貼文的留言",
        "security":[{
          "Bearer":[]
        }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "成功回應",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "array",
                  "example": [{
                    "commenter": {
                      "_id": "627100a67ef69b72689e67f7",
                      "name": "Amy Lin",
                      "avator": "https://imgur.com/m9p1N0H.png"
                    },
                    "content": "哈哈",
                    "createdAt": "2022-05-04T10:20:06.167Z"
                  },
                  {
                    "commenter": {
                      "_id": "627100547ef69b72689e67f6",
                      "name": "Chia Yu",
                      "avator": "https://imgur.com/LwPqBcI.png"
                    },
                    "content": "笑屁喔",
                    "createdAt": "2022-05-05T10:20:06.167Z"
                  }]
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400 (ID 格式錯誤)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/idFormatErrorResponse"
            }
          },
          "400 (查無此 ID)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/notFoundIdErrorResponse"
            }
          },
          "400": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/errorResponse"
            }
          }
        }
      },
      "patch": {
        "tags": ["Post"],
        "description": "新增留言",
        "security":[{
          "Bearer":[]
        }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },{
            "name": "obj",
            "in": "body",
            "description": "資料格式",
            "required":true,
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string",
                  "example": "留言內容"
                }
              },
              "required": [
                "content"
              ]
            },
            "type": "string"
            }
        ],
        "responses": {
          "200": {
            "description": "成功回應",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "array",
                  "example": [{
                    "_id": "62763a35b90092b5356055ab",
                    "author": "627100547ef69b72689e67f6",
                    "content": "更新內容2",
                    "imageUrls": [
                        "https://i.imgur.com/L7EzN1j.png"
                    ],
                    "likes": [],
                    "createdAt": "2022-05-07T09:21:57.477Z",
                    "comments": [
                        {
                            "commenter": "627100547ef69b72689e67f6",
                            "content": "留言",
                            "createdAt": "2022-05-08T09:40:42.081Z",
                            "_id": "6277901ac3db681a968b7c65"
                        }
                    ]
                }]
                },
                "message": {
                  "type": "string",
                  "example": "留言成功"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400 (沒有留言內容)": {
            "description": "失敗回應",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "請輸入留言內容"
                }
              
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400 (ID 格式錯誤)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/idFormatErrorResponse"
            }
          },
          "400 (查無此 ID)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/notFoundIdErrorResponse"
            }
          },
          "400 (尚未登入)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/userIdErrorResponse"
            }
          },
          "400": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/errorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "idFormatErrorResponse":{
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": false
        },
        "message": {
          "type": "string",
          "example": "貼文 ID 格式錯誤，請重新確認"
        }
      
      },
      "xml": {
        "name": "main"
      }
    },
    "userIdErrorResponse":{
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": false
        },
        "message": {
          "type": "string",
          "example": "認證失敗，請重新登入"
        }
      
      },
      "xml": {
        "name": "main"
      }
    },
    "notFoundIdErrorResponse":{
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": false
        },
        "message": {
          "type": "string",
          "example": "查無此篇貼文，貼文已不存在"
        }
      
      },
      "xml": {
        "name": "main"
      }
    },
    "errorResponse":{
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": false
        },
        "message": {
          "type": "string",
          "example": "系統錯誤，請洽系統管理員"
        }
      
      },
      "xml": {
        "name": "main"
      }
    }
  }
}